moved errorUtils to central spot for routes
