+++ backend/routes/spells/index.js
@@ -1,22 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const spellService = require('../../services/entities/spells');
-const ERROR_CODES = require('../../../common/errorCodes');
-
-function mapErrorToStatus(err) {
-  switch (err.code) {
-    case ERROR_CODES.DUPLICATE_ID:
-      return 409;
-    case ERROR_CODES.NOT_FOUND:
-      return 404;
-    case ERROR_CODES.ENTITY_VALIDATION_FAILED:
-    case ERROR_CODES.BUSINESS_LOGIC_FAILED:
-    case ERROR_CODES.INVALID_ID:
-      return 400;
-    default:
-      return 500;
-  }
-}
+const { mapErrorToStatus } = require('../../utils/errorUtils');
 
 // Create a new spell
 router.post('/', async (req, res) => {
